requirements.txt
absl-py==2.4.0
astunparse==1.6.3
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
Flask==3.1.3
flask-cors==6.0.2
flatbuffers==25.12.19
gast==0.7.0
google-pasta==0.2.0
grpcio==1.78.0
h5py==3.14.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
keras==3.13.2
libclang==18.1.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
numpy==2.4.3
opencv-python==4.13.0.92
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pillow==12.1.1
protobuf==7.34.0
Pygments==2.19.2
requests==2.32.5
rich==14.3.3
setuptools==82.0.1
six==1.17.0
tensorflow==2.21.0
termcolor==3.3.0
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
wheel==0.46.3
wrapt==2.1.2
